Extract the item and image operations into separate modules
This should make these operations more reuable and composable. Next
steps will be to implement the image update using the extracted operations.
