+++ lib/melodica_inventory/goods/variation_destroy.ex
@@ -1,5 +1,5 @@
 defmodule MelodicaInventory.Goods.VariationDestroy do
-  alias MelodicaInventory.Goods.ItemDestroy
+  alias MelodicaInventory.Goods.ImageOperations
 
   def build_destroy_action(variation) do
     Ecto.Multi.new
@@ -10,6 +10,6 @@ defmodule MelodicaInventory.Goods.VariationDestroy do
   defp delete_images(items) do
     items
     |> Enum.flat_map(&(&1.images))
-    |> ItemDestroy.delete_images()
+    |> ImageOperations.delete_from_cloudex()
   end
 end
