+++ lib/melodica_inventory_web/templates/admin/item/edit.html.eex
@@ -1,5 +1,5 @@
 <h1><%= @changeset.data.name %></h1>
-<%= form_for @changeset, admin_item_path(@conn, :update, @changeset.data.id), fn f -> %>
+<%= form_for @changeset, admin_item_path(@conn, :update, @changeset.data.id), [multipart: true], fn f -> %>
   <div class="form-group">
     <%= label f, :name %>
     <%= text_input f, :name, type: :text, class: "form-control" %>
@@ -14,6 +14,12 @@
     <%= label f, :price %>
     <%= text_input f, :price, type: :number, class: "form-control", "aria-describedby": "priceHelp", step: "0.01", pattern: "\d*" %>
     <small id="priceHelp" class="form-text text-muted">The price of a single <%= @changeset.data.name %></small>
+  </div>
+  <div class="form-group">
+    <%= label f, :image %>
+    <%= file_input f, :image, class: "form-control" %>
+    <%= error_tag f, :image %>
+    <small class="form-text text-muted">Optional image to replace the currently used one</small>
   </div>
   <%= submit "Submit", class: "btn btn-primary" %>
   <a href="<%= category_path(@conn, :show, @changeset.data.variation.category_id) %>" class="btn btn-secondary">Back</a>
